MA-9021[Android_MX6SL_EVK/MX6SX_SD] Camera/OV5640: Legacy Camera crashes if
switch picture size from 720x480 to VGA. 100%

The reason is that there is a mechanism in LegacyCamera app, which is used
to filter out sizes which aspect ratio is not equal to the picture's aspect
ratio, and then set the max eligible size as preview size. But camera service
can't support modify preview size when camera is preview display.
Remove the mechanism and set the max size as preview size.

Change-Id: I09bcdfafb5a04c406e3a309a08a141fa23fb4990
Signed-off-by: Wenbo.Zhu <wenbo.zhu_1@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/1343
Reviewed-by: Xiaowen Liu <xiaowen.liu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/2990
Reviewed-on: http://androidsource.ap.freescale.net/project/5414